Moved variables to .env
