+++ .gitignore
@@ -70,4 +70,4 @@ static/
 .env
 db.sqlite3
 
-api/views/utils/annular-strata-433816-d1-4df090169a8a.json
+annular-strata-433816-d1-4df090169a8a.json
